change floats to doubles

+++ nebula_decoders/include/nebula_decoders/nebula_decoders_velodyne/decoders/velodyne_generic_decoder.hpp
@@ -65,7 +65,7 @@ public:
 
     // Set up cached values for sin and cos of all the possible headings
     for (uint16_t rot_index = 0; rot_index < ROTATION_MAX_UNITS; ++rot_index) {
-      float rotation = angles::from_degrees(ROTATION_RESOLUTION * rot_index);
+      double rotation = angles::from_degrees(ROTATION_RESOLUTION * rot_index);
       rotation_radians_[rot_index] = rotation;
       cos_rot_table_[rot_index] = cosf(rotation);
       sin_rot_table_[rot_index] = sinf(rotation);
@@ -195,7 +195,7 @@ public:
   void unpack(const velodyne_msgs::msg::VelodynePacket & velodyne_packet)
   {
     const raw_packet_t * raw = (const raw_packet_t *)&velodyne_packet.data[0];
-    float last_azimuth_diff = 0;
+    double last_azimuth_diff = 0;
     uint16_t azimuth_next;
     const uint8_t return_mode = velodyne_packet.data[RETURN_MODE_INDEX];
     const bool dual_return = (return_mode == RETURN_MODE_DUAL);
@@ -228,7 +228,7 @@ public:
           return;  // bad packet: skip the rest
       }
 
-      float azimuth_diff;
+      double azimuth_diff;
       uint16_t azimuth;
 
       // Calculate difference between current and next block's azimuth angle.
@@ -242,7 +242,7 @@ public:
         azimuth_next = raw->blocks[block + (1 + dual_return)].rotation;
 
         // Finds the difference between two successive blocks
-        azimuth_diff = static_cast<float>((36000 + azimuth_next - azimuth) % 36000);
+        azimuth_diff = static_cast<double>((36000 + azimuth_next - azimuth) % 36000);
 
         // This is used when the last block is next to predict rotation amount
         last_azimuth_diff = azimuth_diff;
@@ -308,7 +308,7 @@ public:
             VelodyneLaserCorrection & corrections =
               calibration_configuration_->velodyne_calibration.laser_corrections[laser_number];
 
-            float distance = current_return.uint * SensorT::distance_resolution_m;
+            double distance = current_return.uint * SensorT::distance_resolution_m;
             if (distance > 1e-6) {
               distance += corrections.dist_correction;
             }
@@ -330,23 +330,25 @@ public:
                  (current_block.rotation <= sensor_configuration_->cloud_max_angle * 100 ||
                   current_block.rotation >= sensor_configuration_->cloud_min_angle * 100))) {
                 // convert polar coordinates to Euclidean XYZ.
-                const float cos_vert_angle = corrections.cos_vert_correction;
-                const float sin_vert_angle = corrections.sin_vert_correction;
-                const float cos_rot_correction = corrections.cos_rot_correction;
-                const float sin_rot_correction = corrections.sin_rot_correction;
+                const double cos_vert_angle = corrections.cos_vert_correction;
+                const double sin_vert_angle = corrections.sin_vert_correction;
+                const double cos_rot_correction = corrections.cos_rot_correction;
+                const double sin_rot_correction = corrections.sin_rot_correction;
 
-                const float cos_rot_angle = cos_rot_table_[azimuth_corrected] * cos_rot_correction +
-                                            sin_rot_table_[azimuth_corrected] * sin_rot_correction;
-                const float sin_rot_angle = sin_rot_table_[azimuth_corrected] * cos_rot_correction -
-                                            cos_rot_table_[azimuth_corrected] * sin_rot_correction;
+                const double cos_rot_angle =
+                  cos_rot_table_[azimuth_corrected] * cos_rot_correction +
+                  sin_rot_table_[azimuth_corrected] * sin_rot_correction;
+                const double sin_rot_angle =
+                  sin_rot_table_[azimuth_corrected] * cos_rot_correction -
+                  cos_rot_table_[azimuth_corrected] * sin_rot_correction;
 
                 // Compute the distance in the xy plane (w/o accounting for rotation).
-                const float xy_distance = distance * cos_vert_angle;
+                const double xy_distance = distance * cos_vert_angle;
 
                 // Use standard ROS coordinate system (right-hand rule).
-                const float x_coord = xy_distance * cos_rot_angle;     // velodyne y
-                const float y_coord = -(xy_distance * sin_rot_angle);  // velodyne x
-                const float z_coord = distance * sin_vert_angle;       // velodyne z
+                const double x_coord = xy_distance * cos_rot_angle;     // velodyne y
+                const double y_coord = -(xy_distance * sin_rot_angle);  // velodyne x
+                const double z_coord = distance * sin_vert_angle;       // velodyne z
 
                 const uint8_t intensity = current_block.data[k + 2];
 
@@ -366,8 +368,8 @@ public:
                        other_return.bytes[1] == current_return.bytes[1])) {
                       return_type = static_cast<uint8_t>(drivers::ReturnType::IDENTICAL);
                     } else {
-                      const float other_intensity = block % 2 ? raw->blocks[block - 1].data[k + 2]
-                                                              : raw->blocks[block + 1].data[k + 2];
+                      const double other_intensity = block % 2 ? raw->blocks[block - 1].data[k + 2]
+                                                               : raw->blocks[block + 1].data[k + 2];
 
                       bool first =
                         other_return.uint >=
@@ -434,13 +436,13 @@ private:
   /// @return Resulting flag
 
   // params used by all velodyne decoders
-  float sin_rot_table_[ROTATION_MAX_UNITS];
-  float cos_rot_table_[ROTATION_MAX_UNITS];
-  float rotation_radians_[ROTATION_MAX_UNITS];
+  double sin_rot_table_[ROTATION_MAX_UNITS];
+  double cos_rot_table_[ROTATION_MAX_UNITS];
+  double rotation_radians_[ROTATION_MAX_UNITS];
   int phase_;
   int max_pts_;
   double last_block_timestamp_;
-  std::vector<std::vector<float>> timing_offsets_;
+  std::vector<std::vector<double>> timing_offsets_;
 
   SensorT sensor_;
 };
